requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
barcodenumber==0.5.0
beautifulsoup4==4.12.3
bleach==6.0.0
bleach-allowlist==1.0.3
boto3==1.34.162
botocore==1.34.162
braintree==4.20.0
Brotli==1.1.0
cachetools==4.2.4
cairocffi==1.5.1
certifi==2025.4.26
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.4.2
click==8.2.1
coverage==6.5.0
croniter==2.0.7
cryptography==44.0.3
cssselect==1.3.0
cssselect2==0.8.0
cssutils==2.9.0
dataclasses-json==0.5.14
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
dropbox==11.36.2
email-reply-parser==0.5.12
-e git+https://github.com/frappe/erpnext.git@9febca2981a9ea25fad78a77e0a6d496c6237c53#egg=erpnext
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
Faker==18.10.1
filelock==3.13.4
filetype==1.2.0
fonttools==4.58.4
-e git+https://github.com/frappe/frappe.git@056c65c19c78ec98e0e394ad0ccf376d9995151d#egg=frappe
frappe-bench==5.25.4
freezegun==1.2.2
gitdb==4.0.12
GitPython==3.1.44
gocardless_pro==1.22.0
google-api-core==1.34.1
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.70.0
googlemaps==4.10.0
gunicorn @ git+https://github.com/frappe/gunicorn@bb554053bb87218120d76ab6676af7015680e8b6
hiredis==2.2.3
holidays==0.74
honcho==2.0.0
html5lib==1.1
httplib2==0.22.0
hypothesis==6.77.0
idna==3.10
ipython==8.15.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
ldap3==2.9.1
lxml==5.4.0
markdown2==2.4.13
markdownify==0.14.1
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
maxminddb==2.7.0
maxminddb-geolite2==2018.703
monotonic==1.6
mypy_extensions==1.1.0
num2words==0.5.14
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
parso==0.8.4
passlib==1.7.4
-e git+https://github.com/frappe/payments.git@68b54a752f41e9138af000e0b87915361172095c#egg=payments
paytmchecksum==1.7.0
pdfkit==1.0.0
pexpect==4.9.0
phonenumbers==8.13.55
pickleshare==0.7.5
pillow==11.0.0
plaid-python==7.2.1
ply==3.11
posthog==3.21.0
premailer==3.10.0
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==22.3.5
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.10.0
Pygments==2.19.1
PyJWT==2.8.0
PyMySQL==1.1.1
pyngrok==6.0.0
pyOpenSSL==25.0.0
pyotp==2.8.0
pyparsing==3.2.3
pypdf==3.17.4
pyphen==0.17.2
PyPika==0.48.9
pypng==0.20220715.0
PyQRCode==1.2.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-stdnum==2.1
python-youtube==0.8.3
pytz==2023.3
PyYAML==6.0.2
rapidfuzz==2.15.2
rauth==0.7.3
razorpay==1.4.2
redis==4.5.5
-e git+https://github.com/ArtisanClarinets/repair_portal.git@67db9093c79d5bc6fb769260b15e57a3e31c031b#egg=repair_portal
requests==2.32.4
requests-oauthlib==1.3.1
responses==0.23.1
RestrictedPython==8.0
rq==1.15.1
rsa==4.9.1
s3transfer==0.10.4
semantic-version==2.10.0
sentry-sdk==1.37.1
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.7
sql_metadata==2.11.0
sqlparse==0.5.3
stack-data==0.6.3
stone==3.3.9
stripe==10.12.0
tenacity==8.2.3
terminaltables==3.1.10
tinycss2==1.1.1
tomli==2.0.2
traceback-with-variables==2.0.4
traitlets==5.14.3
types-PyYAML==6.0.12.20250516
typing-inspect==0.9.0
typing_extensions==4.14.0
Unidecode==1.3.8
unittest-xml-reporting==3.2.0
uritemplate==3.0.1
urllib3==2.4.0
uv==0.7.13
vobject==0.9.9
watchdog==3.0.0
wcwidth==0.2.13
weasyprint==59.0
webencodings==0.5.1
Werkzeug==3.0.6
Whoosh==2.7.4
xlrd==2.0.1
zopfli==0.2.3.post1
zxcvbn==4.4.28
